Implémentation du décompte des symptômes et modification de la signature de la méthode dans la classe Launcher pour passer d'un objet Integer à un objet Long dans la Map retournée

+++ Project02Eclipse/src/com/hemebiotech/analytics/OutputFileWriter.java
@@ -9,7 +9,7 @@ public class OutputFileWriter {
      * Write a Map into an output File with sorted elements by natural order (alphabetical)
      * @param mapToWrite map to Write, can have unsorted elements
      */
-    public static void WriteInFile(Map<String,Integer> mapToWrite)
+    public static void WriteInFile(Map<String,Long> mapToWrite)
     {
         //TODO implement FileWriter method
